% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnds_state_contrast.R
\name{.label_region_status_pairwise}
\alias{.label_region_status_pairwise}
\title{Label region_status for standard dndsR pairwise tables using side (query/subject)}
\usage{
.label_region_status_pairwise(d, regions, side = c("query", "subject"))
}
\description{
Requires columns like q_gff_seqname/q_gff_start/q_gff_end or s_* depending on side.
}
\keyword{internal}
