% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R, R/summary.R
\name{.calc_mean_ci}
\alias{.calc_mean_ci}
\title{Compute mean/median/SD/SE/CI for a numeric vector}
\usage{
.calc_mean_ci(x, ci_method = c("normal", "bootstrap"), n_boot = 1000)

.calc_mean_ci(x, ci_method = c("normal", "bootstrap"), n_boot = 1000)
}
\description{
Compute mean/median/SD/SE/CI for a numeric vector

Compute mean/median/SD/SE/CI for a numeric vector
}
\keyword{internal}
