% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_dnds.R
\name{.filter_dnds}
\alias{.filter_dnds}
\title{Filter dNdS annotation table by NA / max_dnds / optional logical expression}
\usage{
.filter_dnds(d, filter_expr = NULL, max_dnds = 10)
}
\description{
Filter dNdS annotation table by NA / max_dnds / optional logical expression
}
\keyword{internal}
