% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnds_contrast.R, R/dnds_state_contrast.R,
%   R/dnds_summary.R
\name{.filter_dnds}
\alias{.filter_dnds}
\title{Filter dNdS annotation table by NA / max_dnds / optional logical expression}
\usage{
.filter_dnds(d, filter_expr = NULL, max_dnds = 10)

.filter_dnds(d, filter_expr = NULL, max_dnds = 10)

.filter_dnds(d, filter_expr = NULL, max_dnds = 10)
}
\description{
Filter dNdS annotation table by NA / max_dnds / optional logical expression

Filter dNdS table by NA / max_dnds / optional expression

Filter dNdS annotation table by NA / max_dnds / optional logical expression
}
\keyword{internal}
