% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_dnds.R
\name{.read_comparisons}
\alias{.read_comparisons}
\title{Read comparison file used throughout dndsR}
\usage{
.read_comparisons(x)
}
\description{
Expects columns: comparison_name, query_fasta, query_gff, subject_fasta, subject_gff
}
\keyword{internal}
