% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_dnds.R
\name{.read_regions_bed}
\alias{.read_regions_bed}
\title{Read and normalize a regions BED-like file}
\usage{
.read_regions_bed(
  regions_bed,
  region_seq_col = NULL,
  region_start_col = NULL,
  region_end_col = NULL,
  region_name_col = NULL
)
}
\description{
Returns a data.frame with columns: seqname, start, end, region_name
}
\keyword{internal}
